Functioning Bot
Fixed motor directions and fixed motor speed for mobile goal lifter when lowering the lift. This makes this iteration of the code the first functioning version of the robot! The cone lifter has not yet been implemented, as it has not yet been built, but the robot can now lift the mobile goal and drive around.